DEV: Added test case for reversed x_value_array and counts input
The test verifies that if both the 'x_value_array' and 'counts' data sets
are input in reversed order (so that the values are decreasing monotonically)
then both arrays are flipped so that the independent variable values meet
the "monotonically increasing" required condition.
In the event that this reversal occurs, a print statement is generated in
order to notify the user that this reversal has occurred.