ENH : added 1D peak refinement
This is a bit of frame work to refine the locations of peaks (or
valleys in principle) of 1D curves.  The assumption is that the
location of the extrema has been identified to pixel(sampling)-level
accuracy and the user has a function that will refine a single peak.

This commit includes
 - peak_refinement
    - logic to loop over all the peaks
    - call refine on each peak
    - marshal the output back to the user
 - two refinement functions