DEV: Changed try/except to appropriate if/elif code commands
While the try/except format that was included originally worked,
using try/except without expecting an exception to be raised
when the try function fails is an inappropriate use of the try
function. Using if/elif statements is more appropriate for what
I am trying to do, which is to selectively populate a metadata
dictionary by reading lines from the original header string.