DEV: Converted try/except functions to more appropriate if/elif
While the original try/except function series worked for the metadata
dictionary population that I was attempting, the use of if/elif
is more appropriate since try/except expects an exception to be raised
if, or when the try function fails. Since I am simply reading the
original header string and populating a new metadata dictionary the
use of if/elif is more appropriate.