API : removed function that was too clever
bin_image_to_1D was far to clever for the benefit it provided.

It alleviated the need to make one call at the cost of adding
a function with took a function as an argument and needed
to be able to blind-pass dictionary of kwargs through.  This
is a nightmare for wrapping at vistrails and difficult to
document or debug for API-usage.