DEV: Revised and added basic image math operations for merge
The implementation that I tried originally did not work due to issues
with wrapping docstrings from numpy. I originally tried importing any
and all functions that are already available in numpy directly through
the API interface, however, param types such as "numpy array" do not
conform to our standard types.
After that I tried to simply reassign the function documentation, but
ran into errors doing that as well, though I will be able to figure that
out after AGU.
SO: Some restructuring remains to be completed but all functionality
works as expected in the current implementation.